HW1/requirements.txt
astroid==3.2.4
blinker==1.8.2
click==8.1.7
debugpy==1.8.5
dill==0.3.8
Flask==3.0.3
Flask-MySQL==1.5.2
greenlet==3.0.3
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
msgpack==1.0.8
mysql-connector-python==9.0.0
neovim==0.3.1
platformdirs==4.3.2
PyMySQL==1.1.1
pynvim==0.5.0
tomlkit==0.13.2
Werkzeug==3.0.4
